+++ packages/projection/src/operators/Mappers/certificates/withStakeKeys.ts
@@ -20,36 +20,34 @@ export interface WithStakeKeys {
  * The intended use case of this operator is to keep track of the current set of active stake keys,
  * ignoring **when** they were registered or unregistered.
  */
-// eslint-disable-next-line sonarjs/cognitive-complexity
 export const withStakeKeys = unifiedProjectorOperator<WithCertificates, WithStakeKeys>((evt) => {
   const register = new Set<Crypto.Hash28ByteBase16>();
   const deregister = new Set<Crypto.Hash28ByteBase16>();
   for (const { certificate } of evt.certificates) {
-    if (Cardano.RegAndDeregCertificateTypes.includes(certificate.__typename as Cardano.RegAndDeregCertificateTypes)) {
-      const {
-        stakeCredential: { hash: stakeCredentialHash }
-      } = certificate as Cardano.RegAndDeregCertificateUnion;
+    if (!Cardano.isCertType(certificate, Cardano.RegAndDeregCertificateTypes)) continue;
 
-      switch (certificate.__typename) {
-        case Cardano.CertificateType.StakeDeregistration:
-        case Cardano.CertificateType.Unregistration:
-          if (register.has(certificate.stakeCredential.hash)) {
-            register.delete(certificate.stakeCredential.hash);
-          } else {
-            deregister.add(certificate.stakeCredential.hash);
-          }
-          break;
-        default:
-          // Stake registration
-          if (deregister.has(stakeCredentialHash)) {
-            deregister.delete(stakeCredentialHash);
-          } else {
-            register.add(stakeCredentialHash);
-          }
-          break;
+    const {
+      stakeCredential: { hash: stakeCredentialHash }
+    } = certificate;
+
+    if (Cardano.isCertType(certificate, Cardano.StakeRegistrationCertificateTypes)) {
+      // Stake registration
+      if (deregister.has(stakeCredentialHash)) {
+        deregister.delete(stakeCredentialHash);
+      } else {
+        register.add(stakeCredentialHash);
       }
+      continue;
+    }
+
+    // Stake deregistration
+    if (register.has(stakeCredentialHash)) {
+      register.delete(stakeCredentialHash);
+    } else {
+      deregister.add(stakeCredentialHash);
     }
   }
+
   const [insert, del] =
     evt.eventType === ChainSyncEventType.RollForward
       ? [[...register], [...deregister]]