+++ packages/wallet/src/services/index.ts
@@ -17,3 +17,4 @@ export * from './AddressDiscovery';
 export * from './HandlesTracker';
 export * from './ChangeAddress';
 export * from './AddressTracker';
+export * from './WalletAssetProvider';