+++ packages/wallet/test/PersonalWallet/load.test.ts
@@ -169,7 +169,7 @@ const assertWalletProperties = async (
   await firstValueFrom(wallet.tip$);
   expect(wallet.tip$.value).toEqual(mocks.ledgerTip);
   // currentEpoch$
-  expect(wallet.currentEpoch$.value?.epochNo).toEqual(currentEpoch.number);
+  expect(wallet.currentEpoch$.value).toEqual(expect.objectContaining({ epochNo: currentEpoch.number }));
   // protocolParameters$
   await firstValueFrom(wallet.protocolParameters$);
   expect(wallet.protocolParameters$.value).toEqual(mocks.protocolParameters);