+++ packages/governance/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.8](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/governance@0.9.7...@cardano-sdk/governance@0.9.8) (2024-02-02)
+
+**Note:** Version bump only for package @cardano-sdk/governance
+
 ## [0.9.7](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/governance@0.9.6...@cardano-sdk/governance@0.9.7) (2024-01-31)
 
 **Note:** Version bump only for package @cardano-sdk/governance