+++ packages/tx-construction/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.17.6](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/tx-construction@0.17.5...@cardano-sdk/tx-construction@0.17.6) (2024-02-02)
+
+### Features
+
+* use new deposit field when building dereg cert ([659f4f0](https://github.com/input-output-hk/cardano-js-sdk/commit/659f4f053ab0ddc9ae9e713e4367dd427008b10c))
+
 ## [0.17.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/tx-construction@0.17.4...@cardano-sdk/tx-construction@0.17.5) (2024-01-31)
 
 **Note:** Version bump only for package @cardano-sdk/tx-construction